fix: unnecessary cast

+++ fauxspark/dist.py
@@ -3,14 +3,14 @@ from typing import Any
 
 
 def uniform(_: Any, n: int) -> np.ndarray:
-    w = np.ones(n).astype(np.float64)
+    w = np.ones(n)
     w /= w.sum()
     return w
 
 
 def zipf(dist: dict[Any, Any], n: int) -> np.ndarray:
     alpha = dist["alpha"]
-    w = np.random.zipf(alpha, n).astype(np.float64)
+    w = np.random.zipf(alpha, n)
     w /= w.sum()
     return w
 
@@ -18,21 +18,21 @@ def zipf(dist: dict[Any, Any], n: int) -> np.ndarray:
 def normal(dist: dict[Any, Any], n: int) -> np.ndarray:
     mu = dist["loc"]
     sigma = dist["scale"]
-    w = np.random.normal(mu, sigma, n).astype(np.float64)
+    w = np.random.normal(mu, sigma, n)
     w /= w.sum()
     return w
 
 
 def pareto(dist: dict[Any, Any], n: int) -> np.ndarray:
     alpha = dist["alpha"]
-    w = np.random.pareto(alpha, n).astype(np.float64)
+    w = np.random.pareto(alpha, n)
     w /= w.sum()
     return w
 
 
 def exponential(dist: dict[Any, Any], n: int) -> np.ndarray:
     scale = dist["scale"]
-    w = np.random.exponential(scale, n).astype(np.float64)
+    w = np.random.exponential(scale, n)
     w /= w.sum()
     return w
